// cfmc-web/src/types/supabase/members.ts
export type ChurchMembers = {
  birthDate?: Date;
  isActive: boolean;
  isBoardMember: boolean;
  isMinisterialTeam: boolean;
  username?: string;
  bio?: string;
  tags?: string[];
  suffix?: string;
  email?: string;
  role: string;
  firstName: string;
  lastName: string;
  phoneNumber?: string;
  isLeader: boolean;
  headline?: string;
  image_banner?: string;
  image_profile?: string;
};